pyproject.toml
[project]
name = "ontokit"
dynamic = ["version"]
description = "Collaborative OWL ontology curation API"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "OntoKit Contributors" }
]
keywords = ["ontology", "owl", "rdf", "semantic-web", "knowledge-graph"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0,<2.12",
    "pydantic-settings>=2.6.0,<2.11",
    "rdflib>=7.1.0",
    "owlready2>=0.47",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "redis>=5.2.0",
    "arq>=0.26.0",
    "httpx>=0.28.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.18",
    "gitpython>=3.1.0",
    "pygit2>=1.13.0",
    "pygithub>=2.5.0",
    "minio>=7.2.0",
    "websockets>=14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/ontokit/ontokit-api"
Documentation = "https://ontokit.github.io/docs"
Repository = "https://github.com/ontokit/ontokit-api"

[project.scripts]
ontokit = "ontokit.runner:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "ontokit/version.py"
pattern = 'VERSION = "(?P<version>[^"]+?)(-dev|-rc)?"'

[tool.hatch.build.targets.wheel]
packages = ["ontokit"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["ontokit"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=ontokit --cov-report=term-missing"
